--- Sat.Recruitment.Api/Controllers/UsersController2.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.IO;
-
-namespace Sat.Recruitment.Api.Controllers
-{
-    public partial class UsersController
-    {
-        private StreamReader ReadUsersFromFile()
-        {
-            var path = Directory.GetCurrentDirectory() + "/Files/Users.txt";
-
-            FileStream fileStream = new FileStream(path, FileMode.Open);
-
-            StreamReader reader = new StreamReader(fileStream);
-            return reader;
-        }
-    }
-}
