+++ Sat.Recruitment.Data/Models/Result.cs
@@ -1,0 +1,8 @@
+﻿namespace Sat.Recruitment.Data.Models
+{
+    public class Result
+    {
+        public bool IsSuccess { get; set; }
+        public string Errors { get; set; }
+    }
+}
