+++ Sat.Recruitment.Data/Models/User.cs
@@ -1,0 +1,12 @@
+﻿namespace Sat.Recruitment.Data.Models
+{
+    public class User
+    {
+        public string Name { get; set; }
+        public string Email { get; set; }
+        public string Address { get; set; }
+        public string Phone { get; set; }
+        public string UserType { get; set; }
+        public decimal Money { get; set; }
+    }
+}
