+++ Sat.Recruitment.Test/UnitTest1.cs
@@ -1,8 +1,3 @@
-using System;
-using System.Dynamic;
-
-using Microsoft.AspNetCore.Mvc;
-
 using Sat.Recruitment.Api.Controllers;
 
 using Xunit;
@@ -15,10 +10,9 @@ namespace Sat.Recruitment.Test
         [Fact]
         public void Test1()
         {
-            var userController = new UsersController();
+            var userController = new UserController();
 
-            var result = userController.CreateUser("Mike", "mike@gmail.com", "Av. Juan G", "+349 1122354215", "Normal", "124").Result;
-
+            var result = userController.CreateUser(new Data.Models.User() { }).Result;//"Mike", "mike@gmail.com", "Av. Juan G", "+349 1122354215", "Normal", "124").Result;
 
             Assert.Equal(true, result.IsSuccess);
             Assert.Equal("User Created", result.Errors);
@@ -27,10 +21,9 @@ namespace Sat.Recruitment.Test
         [Fact]
         public void Test2()
         {
-            var userController = new UsersController();
+            var userController = new UserController();
 
-            var result = userController.CreateUser("Agustina", "Agustina@gmail.com", "Av. Juan G", "+349 1122354215", "Normal", "124").Result;
-
+            var result = userController.CreateUser(new Data.Models.User() { }).Result;//"Agustina", "Agustina@gmail.com", "Av. Juan G", "+349 1122354215", "Normal", "124").Result;
 
             Assert.Equal(false, result.IsSuccess);
             Assert.Equal("The user is duplicated", result.Errors);
